ship: iteration 6 - address final review cleanup
